+++ include/zephyr/drivers/emul_fuel_gauge.h
@@ -55,18 +55,6 @@ __subsystem struct fuel_gauge_emul_driver_api {
  * @retval -EINVAL if mV or mA are 0.
  */
 __syscall int emul_fuel_gauge_set_battery_charging(const struct emul *target, uint32_t uV, int uA);
-static inline int z_impl_emul_fuel_gauge_set_battery_charging(const struct emul *target,
-							      uint32_t uV, int uA)
-{
-	const struct fuel_gauge_emul_driver_api *backend_api =
-		(const struct fuel_gauge_emul_driver_api *)target->backend_api;
-
-	if (backend_api->set_battery_charging == 0) {
-		return -ENOTSUP;
-	}
-
-	return backend_api->set_battery_charging(target, uV, uA);
-}
 
 /**
  * @brief Check if the battery has been cut off.
@@ -78,21 +66,12 @@ static inline int z_impl_emul_fuel_gauge_set_battery_charging(const struct emul 
  * @retval -ENOTSUP if not supported by emulator.
  */
 __syscall int emul_fuel_gauge_is_battery_cutoff(const struct emul *target, bool *cutoff);
-static inline int z_impl_emul_fuel_gauge_is_battery_cutoff(const struct emul *target, bool *cutoff)
-{
-	const struct fuel_gauge_emul_driver_api *backend_api =
-		(const struct fuel_gauge_emul_driver_api *)target->backend_api;
-
-	if (backend_api->is_battery_cutoff == 0) {
-		return -ENOTSUP;
-	}
-	return backend_api->is_battery_cutoff(target, cutoff);
-}
 
 #ifdef __cplusplus
 }
 #endif
 
+#include <zephyr/drivers/fuel_gauge/internal/emul_fuel_gauge_impl.h>
 #include <zephyr/syscalls/emul_fuel_gauge.h>
 
 /**
